fix: varios

+++ back/routes/carritoRoutes.js
@@ -1,30 +1,37 @@
 const express = require("express");
-const { Product, Carrito } = require("../models");
+const { Product, Carrito, Customer } = require("../models");
 const carritoRouter = express.Router();
 
-/* carritoRouter.get("/", async (req, res) => {
-  console.log("FUNCIONA!!");
-  res.send();
+carritoRouter.get("/:id", async (req, res) => {
+  /* Id del customer enviado como param */
+  const { id } = req.params;
+  const customerId = Number(id);
+  
   try {
+    const products = await Carrito.findAll({where: {customerId: customerId}, include: [{model: Product}]})
+    console.log(products);
+    res.send(products)
   } catch (error) {
     console.error(error);
+    res.status(404).send(error);
   }
-}); */
+});
 
 carritoRouter.post("/", async (req, res) => {
-  // req.body { info_producto (cantidad de productos), info_customer }
-
+  /* Deben ser INTEGER */
+  /* Esta ruta también pueden usarla para sumar a la cantidad existente más cantidades, por ejemplo en caso que el usuario clickee varias veces agregar al carrito en el menu principal */
   const { productId, customerId, productQuantity } = req.body;
 
-  /* 
-  SEEDEO!
-  const newProduct = await Product.create({id_api: 16, name: "soy el name", price: 18.9, color: "red", display_size_inch: 6.7, info: "soy la info", general_year: 2020, storage_capacity_gb: 64, cpu_number_of_cores: 8, stock: 8})
-        
-    const newCustomer = await Customer.create({email: "edae@mail.com", password: "1234", full_name: "nombre completo", billing_address: "direccion", default_shipping_address: "direc", country: "pais", phone: "telephone"}) */
+  if (
+    typeof productId !== "number" ||
+    typeof customerId !== "number" ||
+    typeof productQuantity !== "number"
+  )
+    return res.status(400).send("Deben ser INTEGER, no STRING");
 
   try {
     const product = await Product.findByPk(productId);
-    if (!product) return res.status(404).send("NO EXISTE");
+    if (!product) return res.status(404).send("No existe ese usuario en la DB");
 
     const [item, created] = await Carrito.findOrCreate({
       where: { customerId: customerId, productId: productId },
@@ -43,9 +50,9 @@ carritoRouter.post("/", async (req, res) => {
           returning: true,
         }
       );
-      return res.send(updatedItem[0]);
+      return res.status(201).send(updatedItem[0]);
     }
-    res.send(item);
+    res.status(201).send(item);
   } catch (error) {
     console.error(error);
     res.status(404).send(error);
@@ -53,13 +60,18 @@ carritoRouter.post("/", async (req, res) => {
 });
 
 carritoRouter.delete("/", async (req, res) => {
-  // req.body { info_producto, info_customer }
-  const { productId, customerId } = req.query;
+  
+  /* Query con los id, del producto a eliminar y del usuario */
+  /* Para usarlas, se le pega a esta ruta DELETE: "http://localhost:3001/carrito/" y se le agrega "?" con los datos, por ejemplo: "http://localhost:3001/carrito/?productId=2&customerId=1", esto enviaría una variable productId con la string "2", y otra llamada customerId con la string "1" */
 
+  const productId = Number(req.query.productId)
+  const customerId = Number(req.query.customerId)
+  
   try {
     const item = await Carrito.findOne({
       where: { customerId: customerId, productId: productId },
     });
+    if (!item) return res.status(404).send("Relación usuario/producto inexistente")
     await Carrito.destroy({ where: { id: item.dataValues.id } });
     res.send(item);
   } catch (error) {
@@ -69,8 +81,16 @@ carritoRouter.delete("/", async (req, res) => {
 });
 
 carritoRouter.put("/", async (req, res) => {
-  // req.body { info_producto (id, nueva cantidad), info_customer }
+  /* Deben ser INTEGER */
+  /* La cantidad debe ser la NUEVA cantidad del producto */
   const { productId, customerId, productQuantity } = req.body;
+
+  if (
+    typeof productId !== "number" ||
+    typeof customerId !== "number" ||
+    typeof productQuantity !== "number"
+  )
+    return res.status(400).send("Deben ser INTEGER, no STRING");
 
   try {
     const [index, updatedItem] = await Carrito.update(
@@ -80,9 +100,9 @@ carritoRouter.put("/", async (req, res) => {
         returning: true,
       }
     );
-    updatedItem.length
+    index
       ? res.send(updatedItem[0])
-      : res.status(404).send("ITEM NOT FOUND");
+      : res.status(404).send("Relación usuario/producto inexistente");
   } catch (error) {
     console.error(error);
     res.status(404).send(error);
